Updated Othello and NN Inference apps to use timestamps for latency measurements.
NOTE: there is a very subtle bug in the lnic-othello-gpr.c
implementation. If I remove the print statement that prints the
response_cnt and map_cnt then the test fails, but with the print
statement things work fine. I think there may be some lnic_read()
that the compiler is optimizing out, which is causing problems.
Still need to do further investigation.