Adding initial implementation of INT Heavy Hitter prevention program.
Receives special INT reports that indicate flowID, pkt size, timestamp,
and ingress switch IP. The program maintains per-flow state to compute
the average flow rate and compares the result against a threshold value.

If the avg flow rate exceeds the threshold then it sends a msg to the
ingress switch telling it to drop all pkts of the flow.